v1.6.0 (#159)
* fix: do not publish on github as draft

* [150] Show titles only (#158)

* feat: add setting to show only the notes titles

* fix: screenshots env parameter

* fix: ascending sort when taping the tile

* fix: remove "fix ascending sort when taping the tile" from changelogs

* [72] Add F-Droid badge (#163)

* docs: update badges and add F-Droid badge

* docs: add conventional commits badge

* fix: typo in README.md

* [162] Make the search view black in black mode (#164)

* hotfix: only set refresh rate in Android 6 or later

* [141] Choose auto backup folder (#169)

* feat: allow to choose the auto backup directory using SAF

* fix: update info page

* fix: methods order

* [141] fix: fork shared_storage to add custom fixes

* [168] Separate auto export settings (#170)

* [168] feat: separate auto export settings in multiple tiles

* [168] fix: SECURITY.md

* [173] feat: allow to disable only showing the title in the search view (#176)

* [174] feat: allow to focus title when creating a new note (#177)

* [168] Fix auto export (#178)

* [166] fix: auto export frequency preference type

* [166] fix: auto export first time

* [168] fix: auto export on default path

* [168] fix: remove print

* [166] Use type safe routing (#175)

* [166] fix: back action on dialogs

* [166] fix: remove buggy back button interceptors

* [166] feat: use generated type-safe routing

* [166] fix: drawer index

* [166] fix: dialogs

* [166] fix: sort button

* [166] fix: remove print

* [166] fix: settings app bar

* [166] fix: enableOnBackInvokedCallback

* [166] fix: await sheet

* [166] fix: editor page transition

* [165] feat: allow to disable note content preview subdued text (#181)

* [167] feat: allow to set text scaling (#180)

* [60] Add integration tests (#155)

* tests: setup patrol

* fix: notes service get all

* tests: add tests for the app launch and the notes page

* fix: fix ascending sort when taping the tile

* tests: add tests for the app bar of the notes page

* [60] fix: rebase

* [60] fix: tests

* [182] Prepare v1.6.0 (#183)

* [182] docs: add all missing docs

* [182] chore: move translations files

* [182] chore: prepare v1.6.0

* [182] chore: update dependencies

* [182] chore: update localizations

* [182] feat: add Crowdin to the settings

* chore: update changelogs

+++ lib/common/preferences/preferences_utils.dart
@@ -6,6 +6,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 class PreferencesUtils {
   static final PreferencesUtils _singleton = PreferencesUtils._internal();
 
+  /// Default constructor.
   factory PreferencesUtils() {
     return _singleton;
   }
@@ -26,24 +27,30 @@ class PreferencesUtils {
     );
   }
 
-  /// Sets the [key] of the preference to the [value].
+  /// Sets the [preferenceKey] to the [value].
   ///
   /// The type [T] of the value should be a basic type: `bool`, `int`, `double`, `String` or `List<String>`.
-  void set<T>(String key, T value) {
+  Future<void> set<T>(PreferenceKey preferenceKey, T value) async {
+    if (preferenceKey.secure) {
+      _secureStorage.write(key: preferenceKey.name, value: value as String);
+
+      return;
+    }
+
     if (T == dynamic) {
       throw ArgumentError('The type T is required.');
     }
 
     if (T == bool) {
-      _preferences.setBool(key, value as bool);
+      await _preferences.setBool(preferenceKey.name, value as bool);
     } else if (T == int) {
-      _preferences.setInt(key, value as int);
+      await _preferences.setInt(preferenceKey.name, value as int);
     } else if (T == double) {
-      _preferences.setDouble(key, value as double);
+      await _preferences.setDouble(preferenceKey.name, value as double);
     } else if (T == String) {
-      _preferences.setString(key, value as String);
+      await _preferences.setString(preferenceKey.name, value as String);
     } else if (T == List<String>) {
-      _preferences.setStringList(key, value as List<String>);
+      await _preferences.setStringList(preferenceKey.name, value as List<String>);
     }
   }
 
@@ -51,6 +58,10 @@ class PreferencesUtils {
   ///
   /// The type [T] of the value should be a basic type: `bool`, `int`, `double`, `String` or `List<String>`.
   T? get<T>(PreferenceKey preferenceKey) {
+    if (preferenceKey.secure) {
+      throw ArgumentError('The preference is securely stored, use getSecure() instead');
+    }
+
     if (T == dynamic) {
       throw ArgumentError('The type T is required.');
     }
@@ -58,28 +69,29 @@ class PreferencesUtils {
     return _preferences.get(preferenceKey.name) as T?;
   }
 
-  /// Removes the value of the [preferenceKey].
-  Future<void> remove(PreferenceKey preferenceKey) async {
-    await _preferences.remove(preferenceKey.name);
-  }
-
-  /// Clears all the preferences.
-  Future<void> clear() async {
-    await _preferences.clear();
-  }
-
-  /// Securely sets the [preferenceKey] to the [value].
-  void setSecure(PreferenceKey preferenceKey, String value) {
-    _secureStorage.write(key: preferenceKey.name, value: value);
-  }
-
   /// Returns the value of the securely stored [preferenceKey].
   Future<String?> getSecure(PreferenceKey preferenceKey) async {
+    if (!preferenceKey.secure) {
+      throw ArgumentError('The preference is not securely stored, use get<T>() instead');
+    }
+
     return await _secureStorage.read(key: preferenceKey.name);
   }
 
-  /// Deletes the value of the securely stored [preferenceKey].
-  Future<void> deleteSecure(PreferenceKey preferenceKey) async {
-    await _secureStorage.delete(key: preferenceKey.name);
+  /// Removes the value of the [preferenceKey].
+  Future<void> remove(PreferenceKey preferenceKey) async {
+    if (preferenceKey.secure) {
+      await _secureStorage.delete(key: preferenceKey.name);
+
+      return;
+    }
+
+    await _preferences.remove(preferenceKey.name);
+  }
+
+  /// Clears all the preferences and all the securely stored preferences.
+  Future<void> clear() async {
+    await _preferences.clear();
+    await _secureStorage.deleteAll();
   }
 }
